fix (optimisation): extend cache key
From my tests, the request url for the key is not enough to invalidate a cache in all situations. An incorrect token, for example, will cause subsequent GET requests to return an "unauthorised acess" response for the TTL period, even if the token has been corrected. Adding the authorisation headers to the key seems to correct this issue.